LottoTicketMachine, LottoTicketMachineTest 삭제 LottoTickets of 추가 LottoTicketsTest 자동생성 갯수 테스트 추가
